requirements.txt
# Multimodal Dialogue System Requirements

# Core ML/AI Libraries
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
evaluate>=0.4.0
accelerate>=0.20.0

# Computer Vision
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0

# NLP and Text Processing
nltk>=3.7
spacy>=3.4.0
tokenizers>=0.13.0

# Web Interface
streamlit>=1.25.0
streamlit-option-menu>=0.3.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Configuration and Utilities
PyYAML>=6.0
python-dotenv>=0.19.0
tqdm>=4.62.0
click>=8.0.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.950

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Optional: For advanced features
# sentence-transformers>=2.2.0  # For semantic similarity
# gradio>=3.0.0  # Alternative web interface
# fastapi>=0.85.0  # API server
# uvicorn>=0.18.0  # ASGI server
